backend/requirements.txt
anyio==4.4.0
asyncpg>=0.29.0
email-validator>=2.0.0
fastapi==0.115.0
firebase-admin==6.5.0
paho-mqtt==2.1.0
Pillow>=10.4.0
python-dotenv==1.0.1
python-multipart>=0.0.9
requests>=2.31.0
uvicorn[standard]==0.30.6
SQLAlchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
